people-obsidian-2511/requirements.txt
pyobjc-framework-Contacts>=9.0
python-dateutil>=2.8.2
pyyaml>=6.0
